Moved all ApartmentException classes to the new package
