+++ src/main/java/pl/kowalczyk/maciej/spring/learn/api/exception/apartment/ApartmentDeleteException.java
@@ -1,4 +1,4 @@
-package pl.kowalczyk.maciej.spring.learn.api.exception;
+package pl.kowalczyk.maciej.spring.learn.api.exception.apartment;
 
 public class ApartmentDeleteException extends ApartmentException {
