+++ src/main/java/pl/kowalczyk/maciej/spring/learn/service/ApartmentService.java
@@ -4,10 +4,10 @@ import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.OptimisticLockingFailureException;
 import org.springframework.stereotype.Service;
-import pl.kowalczyk.maciej.spring.learn.api.exception.ApartmentCreateException;
-import pl.kowalczyk.maciej.spring.learn.api.exception.ApartmentDeleteException;
-import pl.kowalczyk.maciej.spring.learn.api.exception.ApartmentReadException;
-import pl.kowalczyk.maciej.spring.learn.api.exception.ApartmentUpdateException;
+import pl.kowalczyk.maciej.spring.learn.api.exception.apartment.ApartmentCreateException;
+import pl.kowalczyk.maciej.spring.learn.api.exception.apartment.ApartmentDeleteException;
+import pl.kowalczyk.maciej.spring.learn.api.exception.apartment.ApartmentReadException;
+import pl.kowalczyk.maciej.spring.learn.api.exception.apartment.ApartmentUpdateException;
 import pl.kowalczyk.maciej.spring.learn.repository.ApartmentRepository;
 import pl.kowalczyk.maciej.spring.learn.repository.entity.ApartmentEntity;
 import pl.kowalczyk.maciej.spring.learn.service.mapper.ApartmentMapper;
