+++ src/main/java/pl/kowalczyk/maciej/spring/learn/web/ApartmentController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import pl.kowalczyk.maciej.spring.learn.api.exception.ApartmentException;
+import pl.kowalczyk.maciej.spring.learn.api.exception.apartment.ApartmentException;
 import pl.kowalczyk.maciej.spring.learn.service.ApartmentService;
 import pl.kowalczyk.maciej.spring.learn.web.model.ApartmentModel;
